+++ inst/www/shared/shiny.js
@@ -6659,7 +6659,16 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
   });
 
   $(document).on('keydown', function (e) {
-    if (e.which !== 115 || !e.ctrlKey && !e.metaKey || e.shiftKey || e.altKey) return;
+    if (
+    // if not one of the key combos below
+    !(
+    // cmd/ctrl + fn + f4
+    e.which === 115 && (e.ctrlKey || e.metaKey) && !e.shiftKey && !e.altKey ||
+    // cmd/ctrl + shift + fn + f3
+    e.which === 114 && (e.ctrlKey || e.metaKey) && e.shiftKey && !e.altKey)) {
+      return;
+    }
+
     var url = 'reactlog/mark?w=' + window.escape(exports.shinyapp.config.workerId) + "&s=" + window.escape(exports.shinyapp.config.sessionId);
 
     // send notification
@@ -6672,6 +6681,9 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
         html: html,
         closeButton: true
       });
+    }).fail(function () {
+      // found returned error while marking, should open webpage
+      window.open(url);
     });
 
     e.preventDefault();